python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.121.0",
    "pydantic>=2.12.4",
    "pyjwt>=2.10.1",
]
